Reorganizando app autenticacao
