+++ autenticacao/views/__init__.py
@@ -1,0 +1,3 @@
+from .atualizar_usuario_view import AtualizarUsuarioView
+from .login_view import LoginView
+from .logout_view import LogoutView
